refactor: gmail server_test
