+++ examples/api/gmail/server.py
@@ -1,4 +1,5 @@
 import asyncio
+import logging
 
 from agentkernel.gmail import AgentGmailRequestHandler
 from agentkernel.openai import OpenAIModule
